🚚 Move the TabSettings component to new folder
