+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Tabs/TabSettings.js
@@ -1,0 +1,23 @@
+import ConnectionStatus from '../Components/Settings/ConnectionStatus';
+import CommonSettings from '../Components/Settings/CommonSettings';
+import ExpertSettings from '../Components/Settings/ExpertSettings';
+import SpinnerOverlay from '../../../ReusableComponents/SpinnerOverlay';
+import { SettingsHooks } from '../../../../data';
+
+const TabSettings = () => {
+	const { isReady } = SettingsHooks.useStore();
+
+	if ( ! isReady ) {
+		return <SpinnerOverlay />;
+	}
+
+	return (
+		<div className="ppcp-r-settings">
+			<ConnectionStatus />
+			<CommonSettings />
+			<ExpertSettings />
+		</div>
+	);
+};
+
+export default TabSettings;
