+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Tabs/index.js
@@ -2,7 +2,7 @@ import { __ } from '@wordpress/i18n';
 
 import TabOverview from '../../Overview/TabOverview';
 import TabPaymentMethods from '../../Overview/TabPaymentMethods';
-import TabSettings from '../../Overview/TabSettings';
+import TabSettings from './TabSettings';
 import TabStyling from './TabStyling';
 import TabPayLaterMessaging from '../../Overview/TabPayLaterMessaging';
